chore: 섹션 하위 분리 (#54)
