+++ src/pages/HomeBestDetail/components/sections/BookSummary/ProductDetail.tsx
@@ -1,0 +1,5 @@
+const ProductDetail = () => {
+  return <div>ProductDetail</div>;
+};
+
+export default ProductDetail;
